Factor miniapp helpers into dedicated modules
