+++ tests/api/hello.test.ts
@@ -1,15 +1,16 @@
-import test from 'node:test'
 import assert from 'node:assert/strict'
 
+async function run() {
+  const { GET } = await import('../../app/api/hello/route.ts')
+  const res = await GET(new Request('http://localhost'))
+  assert.equal(res.status, 200)
+  const data = await res.json()
+  assert.deepEqual(data, { message: 'Hello from the API' })
+}
+
 if (typeof Deno !== 'undefined') {
-  // Skip when running under Deno
-  test('skip in deno', () => {})
+  Deno.test('GET /api/hello returns greeting', run)
 } else {
-  test('GET /api/hello returns greeting', async () => {
-    const { GET } = await import('../../app/api/hello/route.ts')
-    const res = await GET()
-    assert.equal(res.status, 200)
-    const data = await res.json()
-    assert.deepEqual(data, { message: 'Hello from the API' })
-  })
+  const { default: test } = await import('node:test')
+  test('GET /api/hello returns greeting', run)
 }
